dnt: Mirror directions of navigation transitions in RTL
This CL plumbs the edge parameter into BackForwardTransitionAnimator so
that it can be used to mirror the back gesture animation. The caller
(manager) is responsible for determining the direction so the animator
is agnostic of RTL - only that it needs to perform an animation from a
given edge.

Also fix the progress computation in 3-button mode to run from 0 to 1
irrespective of the direction, allowing the animator to fully implement
the animation as well as matching the semantics of BackEventCompat (used
by gestural mode).

Bug: 331778964
Change-Id: Iab5ab3a0422ac8c6450f93804f39299df20a64b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5690915
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Jungshik Shin <jshin@chromium.org>
Reviewed-by: Lijin Shen <lazzzis@google.com>
Reviewed-by: Khushal Sagar <khushalsagar@chromium.org>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1331282}