Modified adaptive toolbar bridge to return unknown instead of empty list
This CL modifies the returned value of the JNI
GetRankedSessionVariantButtons method. Currently if the model failed to
execute it returns an empty list, which is inconsistent compared with
its single result equivalent, which returns kUnknown. Now it'll
return a list with a single kUnknown value to simplify refactoring
on its callers.

Bug: b/353311506
Change-Id: I3666e888dd494a68ce07938d5365733d7b74dca6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5731195
Reviewed-by: Patrick Noland <pnoland@chromium.org>
Commit-Queue: Salvador Guerrero Ramos <salg@google.com>
Cr-Commit-Position: refs/heads/main@{#1331315}