Show an error dialog when UpdatePrintSettings() fails
Notify the user with an error dialog if there is an error from
UpdatePrintSettings().

Rename test UpdatePrintSettingsFailsErrorDialog to be
PrintWithPreviewBeforeLoadedUpdatePrintSettingsFails, to reflect that
it is covering the failure scenario where print is initiated before
waiting for the preview to be completely ready.

Bug: 40286396
Change-Id: I3478e6e7f016491443d9fe3b99845d9bf97470f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5730656
Reviewed-by: Lei Zhang <thestig@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Alan Screen <awscreen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1331330}