[OOPPD] Handle terminated service after getting user settings
Crash reports indicate that in some cases the Print Backend service is
terminating after having successfully acquired the user's print
settings from a system dialog.  The browser code needs to account for
this and not crash in such a scenario.

Update the PrintBackendServiceManager to detect this case and not
provide a new client ID for printing when this happens.  The printer
query logic can then provide an error message to notify the user that
there has been a printing failure.

Include some additional cleanup:
- Mark TestPrinterQueryOop::simulate_spooling_memory_errors_ as const.
- Upgrade to CHECK in RegisterPrintDocumentClientReusingClientRemote().
- Fix typo.

Bug: 352469230
Change-Id: I3da7b215b8d4eba2f3ba360190e81487e1ac2391
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5722219
Commit-Queue: Alan Screen <awscreen@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1331344}