[Tab Grid Dialog] Fix assertion failure with binding token
The binding token wasn't being reset in the specific sequence of events
in the bug because the animation to hide the tab grid dialog wasn't
completing before the tab switcher was destroyed. This matters for the
incognito tab switcher where we programmatically switch to the regular
tab model once all the tabs are closed.

The fix for this is to
1) Ensure the dialog hide is triggered when destroying
2) Force any in-flight animations for the tab grid dialog to finish
   if a no-animation hide is requested.

A regression test and unit tests are added.

Fixed: 354745444
Change-Id: Ic5500aa586147c77004d6f45cfbb4d79f3b1acff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5730820
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: Sky Malice <skym@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1331377}