Fix row hover effect after an item is removed from the list.
If the focus is currently on a row and we use the mouse to remove a
different row from the tab list, when there is a single row left in the
tab list, the hover effect and close button is still showing because
the focus is still on that row.

This CL refactors the hover and close button visibility logic
so that hover and close button logic will be consolidated and expose
this logic to the TabListView to force the final row in the list refresh
it's hover and close button visibility when there is a single item left
in the list.

Change-Id: Icfeda5c7acf380ab85792e0dfac9540ab4f8dc5d
Fixed: 354279625
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5727146
Reviewed-by: Alison Gale <agale@chromium.org>
Commit-Queue: Steven Luong <stluong@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1331382}