[Fuchsia] Include fuchsia-ava-nelson into the fuchsia console
The builder is used to run av-analysis related tests in luci
swarming. Now they are running in media lab.

The builder should be green after https://crrev.com/c/5731039.
But I'd like to submit this change first so that the later
change can be tested with tryjobs.

Bug: 40935291
Change-Id: I27b945da2ac1b10d7354d5bd256f759ab08bef61
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5732059
Reviewed-by: Haiyang Pan <hypan@google.com>
Commit-Queue: Zijie He <zijiehe@google.com>
Cr-Commit-Position: refs/heads/main@{#1331400}