[Tab Thumbnails] Ignore possibly stuck captures when preparing tabs
The fact we get stuck waiting on a pending capture seems to be one of
the primary causes of the testToolbarMenuItem_GroupActionAndUndo being
broken.

I suspect that since this is a batched test a previous test in the suite
did not wait for a thumbnail and its capture got stuck pending for a
long time.

Remove these in_flight_captures_ in native more eagerly.
1) Remove the in_flight_captures_ for tabs that get closed.
2) Before a capture iterate over the map and remove any elements where
   the tracker's weak pointer got invalidated.

Bug: 353332369
Change-Id: I508bb9ef13a45c21bc253da3dcafa365e2150c44
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5731599
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Auto-Submit: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1331408}