Fix minor bug in DeskBarViewBase::UpdateDeskIconButtonState().
This method can be called for either the new desk button or
the library button. It should be using the "button" argument
to determine the target bounds instead of "new_desk_button_".

This is just a random bug observed while trying to fix the
main linked bug below.

Bug: b:349872368
Change-Id: I9ae4944b6f9aa2177b1bf7a7f7236a5aafb1f6eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5731677
Commit-Queue: Eric Sum <esum@google.com>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1331409}