[FedCM] Keep track of when choose an account was used
Currently we just pass a parameter to ShowMultiAccountPicker() to tell
it that we want the "Choose an account" title when the relevant button
is clicked. But we also want to preserve the title if we return to this
dialog after choosing an account from that dialog and clicking the back
button. To support this, we keep track of this with a boolean in the
observer.

Fixed: 354756014
Change-Id: Iaa7500350bdc118dadf8c87bd3cd95f6ce5e1557
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5730101
Reviewed-by: Yi Gu <yigu@chromium.org>
Commit-Queue: Yi Gu <yigu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1331462}