DriveSearchProvider: Emit empty results when error occurs
This is a no-op to Launcher, as Launcher gracefully handles the case
where no results are found (for example, when a query does not match any
Drive results when Drive does not error).

This allows clients to immediately know when no Drive results can
possibly be returned.

Bug: b:352425751
Change-Id: I6a6a2100f8f290572aeb263cbc621adafd178634
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5724669
Reviewed-by: CJ Huang <chenjih@google.com>
Commit-Queue: Michael Cui <mlcui@google.com>
Cr-Commit-Position: refs/heads/main@{#1331468}