Remove AttributionReportingRuntimeFeatures
The cross app web attribution reporting feature is shipped and
therefore the check on runtime feature in the network service and the
browser are unnecessary. As there is no short term plan to reuse the
runtime feature delegation implementation, we remove the code to
reduce complexity.

Bug: 352077671
Change-Id: I45f42df1a0a706553aad461c6f0b9ac4b02a58f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5688885
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Nan Lin <linnan@chromium.org>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1331471}