Disable HEVC CBR encondig on Intel platforms with driver issues.
HEVC CBR encoding on Intel platform will break if target bitrate is
updated during encoding. This is not fixed till recent drivers. As a
result, when PlatformHEVCEncoderSupport is enabled, on drivers with flaw
we switch to SW BRC even if MediaFoundationH265UseSWBRC is not set,
as it is the only way to make encoding working.
Also reduce HEVC encoding on Intel to L1T2 when driver with issue is detected.

Bug: 41480904
Change-Id: Ic7eb4bf19af68dfd377714a9fc6869c04f722b38
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5724409
Reviewed-by: Eugene Zemtsov <eugene@chromium.org>
Commit-Queue: Jianlin Qiu <jianlin.qiu@intel.com>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1331510}