Generalize origin checks to include extended app scope
This commit introduces the concept of scoped extensions
within the code that utilizes origin-based checks.
The aim is to ensure consistent behavior for web
applications with scoped extensions across all their
origins.

Instead of checking for origin in
{Isolated}WebAppInstallForceList preference, this change
uses WebAppRegistrar to check if given origin corresponds
to any installed app. This, already existing method in
WebAppRegistrar, is aware of possible scope extensions.

In addition to that, now the check will only return true
if an app was actually successfully installed. Existence
of an origin in force list preference does not guarantee
that app exists in system - the installation might have
failed or still be in progress.

Lastly, this change unifies the check for IWAs and PWAs -
now origins for both app types can be checked in the same
way.

Bug: b/333337342
Change-Id: I100d17cb70b49cec590416eb8116ccea456d9831
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5654911
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Sergii Bykov <sbykov@google.com>
Commit-Queue: Magdalena Skarbińska <mskarbinska@google.com>
Cr-Commit-Position: refs/heads/main@{#1331594}