HistoryBackend: Drop HistoryDBTasks in Closing()
HistoryBackend's destructor ensures (via DCHECK) that there is no
scheduled commit, i.e. basically that Closing() has been called just
before.
This could be violated if there was a posted HistoryDBTask at the time
Closing() was run: The task would keep a refptr to the HistoryBackend,
so it'd delay the destructor invocation.

This CL fixes that by wrapping any current posted HistoryDBTask into
a CancelableOnceClosure, similar to what's done for the
`scheduled_commit_`. That allows us to cancel the task in Closing().

Bug: 354474887
Change-Id: I946d8a2bf90c0a9c124f982ba609425cdeea64e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5724455
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Commit-Queue: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1331602}