Fill SupplementalProfileCache in tests
SupplementalProfileCache was populated when running the full browser,
but not within testing environments. Since SupplementalProfileCache is
where the information on, among others, (E-)AC-3 and xHE-AAC decoding
capability is expected to reside now, [1] tests thought that those
codecs were not supported, even on capable machines. They either had to
be told explicitly, or they were skipped if an IsSupportedAudioType()
guard was present.

We now make it easy for tests to fill SupplementalProfileCache with the
same information that the full browser is seeing. This helps to test
under the same conditions that users see.

[1] df53ebd264c055c079e1b7204099d1356a8c84f0

Bug: 352528975
Change-Id: I3696fa8c23e8b967ea1a7f95205f6e48a78b3471
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5701425
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Wojciech Dzierżanowski <wdzierzanowski@opera.com>
Cr-Commit-Position: refs/heads/main@{#1331673}