Remove use of variable-length array
This changes the boolean arrays used during conversion to be
dynamically allocated rather than placed on the stack, which
is non-standard C++ and triggers a clang warning.

This will likely come with a performance hit, but due to
vector<bool>'s nonstandard implementation, there isn't a good way
around it.

Bug: 349656479
Change-Id: I160f5ee9afd350b2c0b0fee6336d8909238b529c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5731079
Commit-Queue: Devon Loehr <dloehr@google.com>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1331715}