[iOS][LO] Change visibility check condition and add more tests
Visibility check was relying on the tab helper `is_showing_lens_overlay`
which was incidentally correct because only one tab helper is supported.
However the correct approach is to check for the actual UI being
presented on the base view controller.

With this, 5 new tests were added, checking creating, showing and hiding
the overlay

Bug: None
Change-Id: Ie41c17d6ea5324126fb92cbde6ec9c7a8ad5dadc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5730194
Reviewed-by: Christian Xu <christianxu@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Radu Nitescu <radunitescu@google.com>
Cr-Commit-Position: refs/heads/main@{#1331725}