Remove NotFatalUntil::M128 from Aggregation Service CHECKs
These were originally DCHECKs and were upgraded to CHECKs with
NotFatalUntil::M128 in crrev.com/c/5385557. We have not seen these be
triggered in the wild. Additionally, these are now fatal as M128 has
been reached; so, removing the argument should be a no-op.

Bug: 328089064
Change-Id: I287823f2501cc223c16858a282196cddfe4bb979
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5725784
Commit-Queue: Alex Turner <alexmt@chromium.org>
Reviewed-by: Nan Lin <linnan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1331734}