Remove NotFatalUntil::M128 from Private Aggregation CHECKs
These were originally DCHECKs and were upgraded to CHECKs with
NotFatalUntil::M128 in crrev.com/c/5383982. We have not seen these be
triggered in the wild. Additionally, these are now fatal as M128 has
been reached; so, removing the argument should be a no-op.

Bug: 328089064
Change-Id: I5fa0de39c231ab93076dddeb1ea5cd737e858116
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5724514
Reviewed-by: Caleb Raitto <caraitto@chromium.org>
Reviewed-by: Dan McArdle <dmcardle@chromium.org>
Reviewed-by: Yao Xiao <yaoxia@chromium.org>
Commit-Queue: Alex Turner <alexmt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1331760}