Add observer to SetUrl and SetName in Product Specifications multi specific implementation
When SetUrl and SetName were implemented for the multi specifics
implementation, the observers were missed. This CL adds them in.

This is a reland of crrev.com/c/5731545. The reason the test flaked
was because the item level specifics weren't sorted - so the
URLs were effectively returned in a random order. So the test would
pass if the random order was correct and would fail if the random
order was incorrect. I've also added some logging to the matcher
for greater clarity when debugging in future (the error message
listed a stream of bytes which wasn't particularly helpful).

Bug: 354785362
Change-Id: Ib4f5c46ec5f2738d2b55a7b1fc4c20c2828137a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5731329
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Commit-Queue: David Maunder <davidjm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1331800}