Enable custom focus views with custom fields in bubble dialogs.
When we create a tab-modal dialog (bubble), we can include custom
view in addition to the standard buttons. The existing code
would let you specify that the dialog should focus on the custom
view, but the custom view might be an aggregate view (with subviews), and you might actually want the focus to be on a sub view.

The code didn't support that, although it focusable sub views
for the standard fields that could be added to the dialog. This CL
extends that support to custom fields.

Bug: 345847539

Change-Id: Ie22a9b8df4b1ee6a6dcceb9b80723c134cca9339
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5656060
Reviewed-by: Allen Bauer <kylixrd@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1331808}