Pass enterprise endpoints to the ReportingService on updates
Pass ReportingEndpoints enterprise policy-configured endpoints to
the ReportingService after the policy endpoint values are updated.
This allows enterprise endpoints to be eventually added to a pool of
endpoints[1] where enterprise reports are allowed to be sent to. A
future CL will add these endpoints to the ReportingCache.

[1] https://www.w3.org/TR/reporting-1/#endpoint

go/3pcd-enterprise-reporting (Googlers-only)

Bug: 353957526
Change-Id: Id4d1272d70e5ee9f7ef6a030cc011313f4ed47c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5703437
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Reviewed-by: Aaron Selya <selya@google.com>
Reviewed-by: Ayu Ishii <ayui@chromium.org>
Commit-Queue: Fiona Zhang <zfiona@google.com>
Cr-Commit-Position: refs/heads/main@{#1331836}