[A11y] Adds feature to flash screen for notifications
Flashes the screen twice when a notification comes in or
is updated. Adds a flag to hide the feature.

Settings to enable/disable the feature will be added later.
For now, if the flag is on, notifications will flash.

See go/chromeos-flash-screen-demo.

Test: Manual and TODO
Bug: b/341554143
Change-Id: I36f1f0751bfcbd55a430c366130c860301be5e1f
AX-Relnotes: N/A
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5732316
Reviewed-by: Akihiro Ota <akihiroota@chromium.org>
Commit-Queue: Katie Dektar <katie@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1331930}