RenderProcessHost: Expose base::Process::Priority
This gets rid of IsProcessBackgrounded() in favor of
GetPriority(), and changes a lot of its use cases.

In the future, there will be more priority states than just
"foregrounded" and "backgrounded". For now, "foregrounded" maps
to "kUserBlocking" and "backgrounded" maps to "kBestEffort".

The current plans for the additional priority state is to
differentiate between different foreground states, so replacing
"IsProcessBackgrounded()" with "GetPriority() == kBestEffort" is
better than replacing it with "GetPriority() != kUserBlocking".

Bug: 348148939
Change-Id: Ibfc1663489465cba00f13ee3ca49a5f9265cfc5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5713290
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Commit-Queue: Patrick Monette <pmonette@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Reviewed-by: Bo Liu <boliu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1331968}